==> jupyter/1_Search-for-Data-Celiac-Restaurants.ipynb <==
from pymongo import MongoClient
from dotenv import load_dotenv
import os
import requests
import json
from urllib.parse import urlencode, urlparse, parse_qsl
import pickle

# Explore Google API, let's see what we find

db

# We are importing our API key
load_dotenv()

g_key = os.getenv("GOOGLE_KEY2")

# Our first dive into Google API looking for gluten free
data_type = "json"
endpoint = f"https://maps.googleapis.com/maps/api/place/findplacefromtext/{data_type}"
params = {
    
    "input":"gluten-free",
    "inputtype":"textquery",
    "fields":"formatted_address,name,geometry",
    "locationbias":"point:lat,lng",
    "key": f"{g_key}"
}

url_params = urlencode(params)
url = f"{endpoint}?{url_params}"

data = requests.get(url)

# We have a postive response
data

type(data)

data = data.json()

type(data)

# I suppose that because we weren't really specific we didn't get much of a results
# but at least we know it works
print(data)

# Let's get serious here and create a funtcion that we can reuse
def nearby_search(lat, long, radius_in_meters, keyword):
    endpoint = "https://maps.googleapis.com/maps/api/place/nearbysearch/json"
    params = {

        "location":f"{lat},{long}",
        "radius":f"{radius_in_meters}",
        #"rankby" : "<distance - must include keyword,name,or type",
        "keyword":f"{keyword}",
        "inputtype":"textquery",
        "fields":"formatted_address,name,geometry",
        "key": f"{g_key}"
    }

    
    url_params = urlencode(params)
    url = f"{endpoint}?{url_params}"
    data = requests.get(url)
    return data.json()

gluten_free = nearby_search(40.420011, -3.701842, 10_000, "gluten-free")

goog_gf = gluten_free["results"]

# Oh....it's like getting Christmas presents!!! 
len(goog_gf)

# Let's open them and see what we got!
goog_gf[0]

pickle_out = open("/Users/claytonlouden/projects/final_project/pickled/goog_gf.pickle","wb")
pickle.dump(goog_gf, pickle_out)
pickle_out.close()
