==> jupyter/2_Scraping-Web-to-Enrich-Dataset.ipynb <==
import requests
from bs4 import BeautifulSoup
import pandas as pd
import re
import pickle

# First we are going to scrape `findmeglutenfree.com` as `gf`

# This is the way we connect to the webpage we are going to scrape
data_gf = requests.get("https://www.findmeglutenfree.com/es/madrid")

# Response code of 200 is good
data_gf.status_code

# We are now putting the website throught the BS parser
soup = BeautifulSoup(data_gf.text)

# Here we grabbed all the "div" tags to explore what we get
div = soup.select("div")
type(div)

# the 5th Index of this list contains what we are looking for
len(div)

# This is a little more precise but we still don't have exactly what we want
ul = soup.select("ul.list-unstyled.mt-4.mb-4")

# Here is the precise location of the names of our restaruants saved to a variable "name"
name = soup.select("a.align-middle")

# We have to extract the contents contained in these tags and save them
# This is our list of restaurant names
grest = [n.contents for n in name]
len(grest)

# This converts our list of lists to a list of items
gf_rest = [name[0] for name in grest]
gf_rest[:5]   

# Here we have the addresses of out restaurants
address = soup.select("span.sl-addr.mt-2")

# Now they have been extracted we will need to combine the names and the addresses in the database
gaddr = [a.contents for a in address]
len(gaddr)

# This converts our list of lists to a list of items
gf_addr = [name[0] for name in gaddr]
gf_addr[:5]

# Next we are going to scrape `happyceliac.com` as `hc`
**First the name**

# We start the process over again for happyceliac.com
data_hc = requests.get("https://www.happyceliac.com/gluten-free-madrid/")

data_hc

soup = BeautifulSoup(data_hc.text)

soup

name = soup.select("h3")

# We had one name at the end of the list that was still nesting bewteen tags 
# We are extracting it from the tags here
rest = [n.contents for n in name]
rest = rest[:-9]
r = [n.contents for n in rest[-1]]

# We are replacing that extracted item back into the list where is was
rest[-1]=r[0]
rest[-5:]

# We create a list of items
hc_rest = []
for n in rest:
    line = re.sub(r"^\s|\s$|\s[★]", "", n[0])
    hc_rest.append(line)
hc_rest

**Now the Address**

address = soup.select("em")
address

# There were some unwanted items in this list so we selected the index of the desired lines
addr = [a.contents for a in address[1:-9]]
addr

# and we selected the first index of each list-type item because that contained the address
new = [line[0] for line in addr]
new[:5]

# Here we cleaned the addresses and they look good
hc_addr = []
for l in new:
    line = re.sub(r"\s[|]\s.*", "", l)
    hc_addr.append(line)

print(hc_addr[:5])
len(hc_addr)

# We are going to scrape the `viajarsingluten` as `vsg`

vsg_rest_tot = []
vsg_addr_tot = []

# These are all the next page links
url_list = ["https://www.viajarsingluten.com/restaurantes-sin-gluten-madrid-1_6047/",
            "https://www.viajarsingluten.com/restaurantes-sin-gluten-madrid-1_6047/2/",
            "https://www.viajarsingluten.com/restaurantes-sin-gluten-madrid-1_6047/3/",
            "https://www.viajarsingluten.com/restaurantes-sin-gluten-madrid-1_6047/4/"]


# This loops through each url and grabs what we need and converts it to a list
for url in url_list:
    # Empty nested lists 
    vsg_rest = []
    vsg_addr = []

    # Get url
    print(url)
    data_vsg = requests.get(url)
    soup = BeautifulSoup(data_vsg.text)
    
    # Select specific tags
    names = soup.select("div.nombre")
    address = soup.select("div.dir")
    
    # Select "a" tags within "nombre" tags
    name = [n.select("a") for n in names]

    # Extract contents from tags
    vrest = [n[0].contents for n in name]
    vaddr = [a.contents for a in address]

    # Create a list of items from list of lists
    vsg_rest = [name[0] for name in vrest] 
    vsg_addr = [name[0] for name in vaddr]
    
    # Append alll findings to master list
    vsg_rest_tot.append(vsg_rest)
    vsg_addr_tot.append(vsg_addr)       

# That worked! We need to flatten this list of lists.  
# Not a very elegant solution but...quick...while no one's looking! 
vsg_flat_rest = [y for x in vsg_rest_tot for y in x]
vsg_flat_rest[20:30]

# There we have eliminated that space in case it causes us problems when searching in Google's API
for index, name in enumerate(vsg_flat_rest): 
    clean = re.sub(r"^\s", "", name)
    vsg_flat_rest[index] = clean    

vsg_flat_rest[20:30]

# Now do the same with the addresses: flatten and trim that extra white space
vsg_flat_addr = [y for x in vsg_addr_tot for y in x]
vsg_flat_addr[30:35]

for index, addr in enumerate(vsg_flat_addr): 
    clean = re.sub(r"^\s", "", addr)
    vsg_flat_addr[index] = clean    

vsg_flat_addr[30:35]

# Finally, we need to combine these lists into dictionaries

# These are lists of items
gf_rest
gf_addr
hc_rest
hc_addr

# These are much longer lists that we might use
vsg_flat_rest
vsg_flat_addr

# Lets let's make a function to do the work for us.
gf_list = []

def add_n_a_lists_to_dict(namelist, addresslist, list_dict):
    results = set(zip(namelist,addresslist))
    for tup in results:
        dic = {}
        dic["name"] = tup[0]
        dic["address"] = tup[1]
        list_dict.append(dic)
    return list_dict

add_n_a_lists_to_dict(gf_rest,gf_addr, gf_list)
add_n_a_lists_to_dict(hc_rest,hc_addr, gf_list)

first_places = gf_list

len(first_places)

first_places[0]


list(first_places[1].values())

" ".join(list(first_places[1].values()))

gf_list = []

def add_n_a_lists_to_dict(namelist, addresslist, list_dict):
    results = set(zip(namelist,addresslist))
    for tup in results:
        dic = {}
        dic["name"] = tup[0]
        dic["address"] = tup[1]
        list_dict.append(dic)
    return list_dict

second_places = add_n_a_lists_to_dict(vsg_flat_rest,vsg_flat_addr, gf_list)

len(second_places)

%store first_places
%store second_places
# ERROR: maximum recursion depth exceeded while pickling an object

# The %store magic works fine here
from sklearn import datasets
dataset = datasets.load_iris()
%store dataset

-------------

# For some reason pickling didn't work with `first_places` so we cut and paste... 

pickle_out = open("/Users/claytonlouden/projects/final_project/pickled/first_places.pickle","wb")
pickle.dump(first_places, pickle_out)
pickle_out.close()
# maximum recursion depth exceeded while calling a Python object

# Save `second_places` 

pickle_out = open("/Users/claytonlouden/projects/final_project/pickled/second_places.pickle","wb")
pickle.dump(second_places, pickle_out)
pickle_out.close()

-----------
