==> jupyter/3_Enrich-Our-Google-Results.ipynb <==
import os
from dotenv import load_dotenv
import requests
import json
from urllib.parse import urlencode, urlparse, parse_qsl
import pickle

# Now we take our Web Scrapings and pass them through Google's API

load_dotenv()

g_key = os.getenv("GOOGLE_KEY2")

# This is a test to get an idea of what our parameters should look like and what info is returned
endpoint = "https://maps.googleapis.com/maps/api/place/findplacefromtext/json"
params = {

    "input":f"{New York Burger Calle del Gral Yagüe, 5, 28020 Madrid, Madrid, Spain}",
    "inputtype":"textquery",
    "fields":"place_id,name,formatted_address,geometry,business_status,photos,price_level,rating,user_ratings_total,opening_hours,plus_code,types,icon",
    "key": f"{g_key}"
}


url_params = urlencode(params)
url = f"{endpoint}?{url_params}"


data = requests.get(url)

data

type(data)

data = data.json()

type(data)

# We have some our first results
data

------------

# Now we are going to import our variables from our previous Jupyter:

----------

# For some reason pickling didnt work with `first_places` so we cut and paste... 


pickle_in = open("/Users/claytonlouden/projects/final_project/pickled/first_places.pickle","rb")
first_places = pickle.load(pickle_in)

first_places = [{'name': 'New York Burger', 'address': 'Calle del Gral Yagüe, 5, 28020 Madrid, Madrid, Spain'}, {'name': "Alfredo's Barbacoa Restaurant", 'address': 'Calle de Juan Hurtado de Mendoza, 11, 28036 Madrid, Madrid, Spain'}, {'name': 'Flavia Madrid', 'address': 'Calle Gil de Santivañes, 2, 28001 Madrid, Spain'}, {'name': 'El Senador', 'address': 'Plaza de la Marina Española, 2, 28013 Madrid, Madrid, Spain'}, {'name': 'Restaurante lapizza+sana', 'address': 'Av. de Machupichu, 16, 28043 Madrid, Spain'}, {'name': 'Oriental Pastry', 'address': 'Calle Ferraz, 47, 28008 Madrid, Spain'}, {'name': 'Taberna La Concha', 'address': 'Calle Cava Baja, 7, 28005 Madrid, Spain'}, {'name': 'As De Bastos Madrid', 'address': 'Calle de Castilla, 62, 28039 Madrid, Spain'}, {'name': 'Maestro Churrero', 'address': 'Plaza de Jacinto Benavente, 2, 28012 Madrid, Spain'}, {'name': 'El Anciano Rey de los Vinos', 'address': 'Calle de Bailén, 19, 28013 Madrid, Madrid, Spain'}, {'name': 'Kint', 'address': 'Calle Vallehermoso, 36, 28015 Madrid, Spain'}, {'name': 'Da Nicola Gran Vía', 'address': 'Plaza de los Mostenses, 11, 28015 Madrid, Community of Madrid, Spain'}, {'name': 'rodilla', 'address': 'Plaza Puerta del Sol, 13, 28013 Madrid, Madrid, Spain'}, {'name': 'Celicioso', 'address': 'Calle Hortaleza, 3, Madrid, Spain'}, {'name': 'II Piccolino della Farfalla', 'address': 'Calle de Las Huertas, 6, Madrid, Spain'}, {'name': 'Solo de Croquetas', 'address': 'Calle de Echegaray, 5, 28014 Madrid, Spain'}, {'name': 'Emma Y Julia', 'address': 'Calle Cava Baja, 19, 28005 Madrid, Madrid, Spain'}, {'name': 'La Jauría', 'address': 'Calle de Colón, 10, 28004 Madrid, Spain'}, {'name': 'El Arrozal', 'address': 'Calle Segovia, 13, 28005 Madrid, Spain'}, {'name': 'Sana Locura Gluten Free Bakery', 'address': 'Calle del Gral Oraá, 49, 28006 Madrid, Spain'}, {'name': 'La Mordida', 'address': 'Av. de Brasil, 6, 28020 Madrid, Spain'}, {'name': 'Rosa Negra Madrid', 'address': 'Calle del León, 16, 28014 Madrid, Madrid, Spain'}, {'name': 'Mercado de la Reina', 'address': 'Calle Gran Vía, 12, 28013 Madrid, Spain'}, {'name': 'LAIB', 'address': 'Paseo de la R. Cristina, 30, 28014 Madrid, Spain'}, {'name': 'Celicioso Barquillo', 'address': 'Calle del Barquillo, 19, 28004 Madrid, Spain'}, {'name': 'Murillo Café', 'address': 'Calle Ruiz de Alarcón, 27, 28014 Madrid, Madrid, Spain'}, {'name': 'Restaurante Vegetariano Artemisa Sol', 'address': 'Calle de las Tres Cruces, 4, 28013 Madrid, Spain'}, {'name': 'El Corte Inglés', 'address': 'Calle Preciados, 3, 28013 Madrid, Spain'}, {'name': 'Okashi Sanda', 'address': 'Calle de San Vicente Ferrer, 22, 28004 Madrid, Spain'}, {'name': 'CafeeKe', 'address': 'Calle de Cuchilleros, 3, 28005 Madrid, Spain'}, {'name': 'Celicioso', 'address': 'Various Locations'}, {'name': 'Bico de Xeado', 'address': 'Various Locations'}, {'name': 'As De Bastos', 'address': 'Calle de Castilla, 62, 28039 Madrid, Spain'}, {'name': 'La Oriental Sin Gluten', 'address': 'Calle de Ferraz, 47, 28008 Madrid, Spain'}, {'name': 'Maestro Churro', 'address': 'Plaza de Jacinto Benavente, 2, 28012 Madrid, Spain'}, {'name': 'Celikatessen', 'address': 'Mercado de Pacífico (planta baja, Calle de Valderribas, 30, 28007 Madrid, Spain)'}, {'name': 'El Arrozal', 'address': 'Calle de Segovia, 13, 28005 Madrid, Spain'}, {'name': 'Okashi Sanda', 'address': 'Calle de San Vicente Ferrer, 22, 28004 Madrid, Spain'}, {'name': 'Artemisa Huertas', 'address': 'Calle de Ventura de la Vega, 4, 28014 Madrid, Spain'}, {'name': 'Sana Local', 'address': 'Calle del Gral Oraá, 49, 28006 Madrid, Spain'}, {'name': 'El Antiguo: Cafetería y Tapas', 'address': 'Avenida de Madrid, 77, 28607 El Álamo, Spain'}, {'name': 'Asador Maribel', 'address': 'Av. del Padre Claret, 16, 40001 Segovia, Spain'}, {'name': 'La Taberna Concha', 'address': 'Calle de la Cava Baja, 7, 28005 Madrid, Spain'}, {'name': 'Da Nicola', 'address': 'Plaza Mostenses, 11, 28015 Madrid, Spain'}, {'name': 'Loliam', 'address': 'Calle Villalar, 6, 28001 Madrid, Spain'}, {'name': 'Bar Mentrida', 'address': 'Plaza de Olavide, 3, 28010 Madrid, Spain'}, {'name': 'Kint Bistro', 'address': 'Calle de Vallehermoso, 36, 28015 Madrid, Spain'}, {'name': 'Solo de Croquetas', 'address': 'Calle de Echegaray, 5, 28014 Madrid, Spain'}]

first_places[:5]

---------

# but it worked for the much longer list of dictionaries `second_places`
# Retrieve `second_places` 

pickle_in = open("/Users/claytonlouden/projects/final_project/pickled/second_places.pickle","rb")
second_places = pickle.load(pickle_in)

second_places[:5]

-------

# This is the function that will take all our requests to Google API and give us back info 
def get_info(inp):
    endpoint = "https://maps.googleapis.com/maps/api/place/findplacefromtext/json"
    params = {

        "input":f"{inp}",
        "inputtype":"textquery",
        "fields":"place_id,name,formatted_address,geometry,business_status,photos,price_level,rating,user_ratings_total,opening_hours,plus_code,types,icon",
        "key": f"{g_key}"
    }


    url_params = urlencode(params)
    url = f"{endpoint}?{url_params}"
    data = requests.get(url)
    return data.json()

# We will create an empty list here
info_list = []

# This is our loop that will take each dictionary item and turn the values into one string
# Then it will pass it through our get_info function and append the output to our empty info_list
for pl in first_places:
    inp = " ".join(list(pl.values()))
    #print(inp)
    info = get_info(inp)
    info_list.append(info)


second_info_list = []

for pl in second_places:
    inp = " ".join(list(pl.values()))
    #print(inp)
    info = get_info(inp)
    second_info_list.append(info)

# first Google API Request results
len(info_list)

# second Google API Requests results
len(second_info_list)

------------------------

# Save `info_list` and `second_info_list`

pickle_out = open("/Users/claytonlouden/projects/final_project/pickled/info_list.pickle","wb")
pickle.dump(info_list, pickle_out)
pickle_out.close()

pickle_out = open("/Users/claytonlouden/projects/final_project/pickled/second_info_list.pickle","wb")
pickle.dump(second_info_list, pickle_out)
pickle_out.close()

--------------
