==> jupyter/4_Entering_Data_in_Database.ipynb <==
from pymongo import MongoClient
from dotenv import load_dotenv
import pickle
import pandas as pd
import os
from urllib.parse import urlencode, urlparse, parse_qsl
import requests

# Now we are going to place this into a Pandas DataFrame

# This is our database connection
client = MongoClient("mongodb://localhost")
db = client.get_database("celiacdb")

db

------

# *We have pickled our data and will import it here*

# This was the original API request we made
pickle_in = open("/Users/claytonlouden/projects/final_project/pickled/goog_gf.pickle","rb")
goog_info = pickle.load(pickle_in)

len(goog_info)

# These are the results from findmeglutenfree.com and happyceliac.com 
pickle_in = open("/Users/claytonlouden/projects/final_project/pickled/info_list.pickle","rb")
first_info = pickle.load(pickle_in)

len(first_info)

# These are the results from viajarsingluten.com
pickle_in = open("/Users/claytonlouden/projects/final_project/pickled/second_info_list.pickle","rb")
second_info = pickle.load(pickle_in)

len(second_info)

------

### *There's a little more clean up we need to do*

# Looking for number of key:value pairs by looking at length
# We also have observed that we need to flatten first_info and second_info
for pl in second_info[:5]:
    for p in pl['candidates']:
        print(len(p))
        

# We are going to put the contents into a dummy variable so that we can put them back into the original variable
second = second_info

# This is a single key:value pair. We only want to extract the value from the key "candidates"
second[0]

# We empty the orginial variable and it is ready to be filled with the dictionaries
second_info = []

# Now this is going to append the list in the original variable with just what we need
for pl in second:
    for p in pl['candidates']:
        second_info.append(p)

# Perfect
second_info[0]

first_info[0]

goog_info[8]

first = first_info

first_info = []

for pl in first:
    for p in pl['candidates']:
        first_info.append(p)

# perfect here too
first_info[6]

load_dotenv()

g_key = os.getenv("GOOGLE_KEY2")

endpoint = "https://maps.googleapis.com/maps/api/place/photo"
params = {
    
    "photoreference" : "ATtYBwJDdX67hPxKnqZLWc2w92JIO6QYQEP8ESIXnE6VSXO-gU_4VZymWj44I7h-t_vy8hlyNHaP1PMs8MOzBPSzXcbPADvWq_vZO4D4K5pWRWmwIld0gNzWYOk7mxhpN180PGStPRkO6lTOgks2VFb0ScXMNkUGHIirzqqVbVwi1mf7Hx8s",
    "maxwidth" : "400",
    "maxheight" : "600",
    "key" : f"{g_key}"
}
url_params = urlencode(params)
url = f"{endpoint}?{url_params}"
data = requests.get(url)

data

### *We are going to make all the dictionaries contain the same keys*

first_info[0].keys() == second_info[0].keys() == goog_info[0].keys()

all_data = goog_info + first_info + second_info


len(all_data)

all_data[150]

all_data[0]["photos"][0]['photo_reference']

needed_info = ["geometry","name","place_id","price_level","rating","types","user_ratings_total","vicinity",'photos']

all_data_same_keys = []

for tienda in all_data:
    temp_dict = {}
    for y in needed_info:
        if y not in list(tienda.keys()):
            tienda[y] = "Unknown"
        if y == "geometry":
            temp_dict[y]= tuple(tienda[y]["location"].values())
        
        else:
            temp_dict[y] = tienda[y]
        
    all_data_same_keys.append(temp_dict)
        

len(all_data_same_keys)

all_data_same_keys[0]

for tienda in all_data_same_keys:
    try:
        tienda["photos"] = tienda["photos"][0]["photo_reference"]
    except:
        print(tienda["photos"])
    

all_data_same_keys[0]

### *We have to clean the `types` category*

# We will only label restaruant, bakery, bar, cafe, store, supermarket, or locale

for d in all_data_same_keys:
    if "bakery" in d["types"] and "restaurant" in d["types"]:
        d["types"] = "restaurant/bakery"
    elif "bakery" in d["types"]:
        d["types"] = "bakery"
    elif "restaurant" in d["types"]:        
        d["types"] = "restaurant"
    elif "cafe" in d["types"]:
        d["types"] = "cafe"
    elif "store" in d["types"]:
        d["types"] = "store"
    elif "supermarket" in d["types"]:
        d["types"] = "supermarket"
    elif "lodging" in d["types"]:
        d["types"] = "locale"
    else:
        d["types"] = "locale"

### *Now we can get back to creating our DataFrame*

all_data_same_keys[0]

df = pd.DataFrame(data=all_data_same_keys)

df.head()

df.place_id.value_counts().head()

df.drop_duplicates(subset=["place_id"],ignore_index=True,inplace=True)

df.place_id.value_counts().head()

df.columns

df = df[['name','rating','user_ratings_total','price_level','types','vicinity','geometry','place_id',
       'photos']]

df.name.value_counts().head()

chains = ["VIPS","Rodilla", "Ginos", "McDonald's", "Telepizza - Comida a domicilio", "Telepizza", "Tommy Mel's", "Fridays", "Pizza Jardín"]


ch = []
for n in df.name:
    if n in chains:
        ch.append("chain")
    else:
        ch.append("unique")


df["chain"] = ch

df = df[['name','rating','user_ratings_total','price_level','types','chain', 'vicinity','geometry','place_id',
       'photos']]

df.head(50)

---------

# Save `df` as `gfdf`

pickle_out = open("/Users/claytonlouden/projects/final_project/pickled/gfdf.pickle","wb")
pickle.dump(df, pickle_out)
pickle_out.close()
